test: move mockReactVisibilitySensor to shared/testing
