+++ client/branded/src/search-ui/components/FileMatchChildren.test.tsx
@@ -10,7 +10,7 @@ import {
     NOOP_SETTINGS_CASCADE,
     HIGHLIGHTED_FILE_LINES,
 } from '@sourcegraph/shared/src/testing/searchTestHelpers'
-import '@sourcegraph/shared/dev/mockReactVisibilitySensor'
+import '@sourcegraph/shared/src/testing/mockReactVisibilitySensor'
 import { renderWithBrandedContext } from '@sourcegraph/wildcard/src/testing'
 
 import { FileMatchChildren } from './FileMatchChildren'
