+++ client/web/src/codeintel/ReferencesPanel.test.tsx
@@ -2,7 +2,7 @@ import { within, fireEvent } from '@testing-library/react'
 import { createPath } from 'react-router-dom'
 
 import { MockedTestProvider, waitForNextApolloResponse } from '@sourcegraph/shared/src/testing/apollo'
-import '@sourcegraph/shared/dev/mockReactVisibilitySensor'
+import '@sourcegraph/shared/src/testing/mockReactVisibilitySensor'
 import { renderWithBrandedContext } from '@sourcegraph/wildcard/src/testing'
 
 import { setExperimentalFeaturesFromSettings } from '../stores'
